Errors gefixd

+++ DAL/MenuItemDao.cs
@@ -58,7 +58,7 @@ namespace DAL
                 {
                     MenuItemId = Convert.ToInt32(dr["MenuItemId"]),
                     Voorraad = Convert.ToInt32(dr["Voorraad"]),
-                    Prijs = (decimal)dr["Prijs"],
+                    Prijs = (double)dr["Prijs"],
                     Naam = (string)dr["Naam"],
                     MenuId = Convert.ToInt32(dr["MenuId"]),
                     IsAlcoholisch = (bool)dr["Alcoholisch"],
