+++ Model/MenuItem.cs
@@ -14,7 +14,7 @@
 
         }
 
-        public MenuItem(string naam, decimal prijs, bool isAlcoholisch, int menuId, int voorraad)
+        public MenuItem(string naam, double prijs, bool isAlcoholisch, int menuId, int voorraad)
         {
             Naam = naam;
             Prijs = prijs;
